chore: move MemoryBytesFlag to facade
Use it for memory related flags
Also change pipeline_squash to 1 by default.
